+++ alu-web/src/notes/editor-components/shortcuts.tsx
@@ -1,0 +1,111 @@
+import {
+  Editor,
+  Transforms,
+  Range,
+  Point,
+  Element as SlateElement,
+} from 'slate';
+
+const SHORTCUTS = {
+  '*':      { type: 'list-item', extra: 'bulleted-list' },
+  '-':      { type: 'list-item', extra: 'bulleted-list' },
+  '+':      { type: 'list-item', extra: 'bulleted-list' },
+  '1.':     { type: 'list-item', extra: 'numbered-list' },
+  '$$':     { type: 'math-block' },
+  '#':      { type: 'heading-one' },
+  '##':     { type: 'heading-two' },
+  '###':    { type: 'heading-three' },
+  '####':   { type: 'heading-four' },
+  '#####':  { type: 'heading-five' },
+  '######': { type: 'heading-six' },
+};
+
+// Taken from https://github.com/ianstormtaylor/slate/blob/cb1db7fca914854a61c8f05dca416500d18ce7d3/site/examples/markdown-shortcuts.tsx
+export const withShortcuts = editor => {
+  const { deleteBackward, insertText } = editor;
+
+  editor.insertText = text => {
+    const { selection } = editor;
+    if (text === ' ' && selection && Range.isCollapsed(selection)) {
+      const { anchor } = selection;const block = Editor.above(editor, {
+        match: n => Editor.isBlock(editor, n),
+      });
+      const path = block ? block[1] : [];
+      const start = Editor.start(editor, path);
+      const range = { anchor, focus: start };
+      const beforeText = Editor.string(editor, range);
+      const { type, extra } = SHORTCUTS[beforeText];
+
+      if (type) {
+        Transforms.select(editor, range);
+        Transforms.delete(editor);
+        const newProperties: Partial<SlateElement> = {
+          type,
+        };
+        Transforms.setNodes(editor, newProperties, {
+          match: n => Editor.isBlock(editor, n),
+        });
+
+        if (type === 'list-item') {
+          const list = {
+            type: extra,
+            children: [],
+          };
+          Transforms.wrapNodes(editor, list, {
+            match: n =>
+              !Editor.isEditor(n) &&
+              SlateElement.isElement(n) &&
+              n.type === 'list-item',
+          });
+        }
+
+        return;
+      }
+    }
+
+    insertText(text);
+  }
+
+  editor.deleteBackward = (...args) => {
+    const { selection } = editor;
+
+    if (selection && Range.isCollapsed(selection)) {
+      const match = Editor.above(editor, {
+        match: n => Editor.isBlock(editor, n),
+      });
+
+      if (match) {
+        const [block, path] = match;
+        const start = Editor.start(editor, path);
+
+        if (
+          !Editor.isEditor(block) &&
+          SlateElement.isElement(block) &&
+          block.type !== 'paragraph' &&
+          Point.equals(selection.anchor, start)
+        ) {
+          const newProperties: Partial<SlateElement> = {
+            type: 'paragraph',
+          };
+          Transforms.setNodes(editor, newProperties);
+
+          if (block.type === 'list-item') {
+            Transforms.unwrapNodes(editor, {
+              match: n =>
+                !Editor.isEditor(n) &&
+                SlateElement.isElement(n) &&
+                (n.type === 'bulleted-list' || n.type === 'numbered-list'),
+              split: true,
+            });
+          }
+
+          return;
+        }
+      }
+
+      deleteBackward(...args);
+    }
+  }
+
+  return editor;
+}