+++ pages/markdown/misc/changelog.md
@@ -11,6 +11,7 @@
 * **Improved Autocomplete for Assignments:** Improved autocomplete for when teachers are creating assignments.  If the assignment title contains "essential" in it, Alu will now automatically fill the tag query to be: "Unit # AND essential".
 * **Unfriend Confirmation:** Added confirmation before removing a user as a friend, to prevent accidentally unfriending someone.
 * **Removed Unnecessary API Request:** Fixed a bug that caused Alu to look for a list of your friends every time you opened the homepage, rather than just when you clicked the friends dropdown.  This will marginally increase loading speeds.
+* **Added Formatting Shortcuts:** When writing flashcards/notes, you can now use shortcuts to automatically format the text.  You can use "*", "-", or "+" to automatically create a bulleted list, "1." to create a numbered list, "$$" to create a math block, and 1-6 "#"s to enter that level of heading.
 
 
 ## Overflow Bucket - May 03 - 0.6.3